+++ pa_dlna/__init__.py
@@ -3,7 +3,7 @@
 import sys
 import logging
 
-__version__ = '0.16'
+__version__ = '1.0'
 MIN_PYTHON_VERSION = (3, 8)
 MIN_LIBPULSE_VERSION = '0.7'